apds9960

Struct Apds9960

pub struct Apds9960<I2C, M: Mode> { /* private fields */ }

APDS9960 device driver.

Implementations

impl<I2C, E> Apds9960<I2C, SyncNonBlocking>
where I2C: I2c<Error = E>,

Common configuration.

pub fn enable(&mut self) -> Result<(), Error<E>>

Turn power on.

pub fn disable(&mut self) -> Result<(), Error<E>>

Deactivate everything and put the device to sleep.

pub fn enable_wait(&mut self) -> Result<(), Error<E>>

Enable the wait feature.

Enables delay between proximity and / or color and ambient light cycles. The duration of the wait can be configured with set_wait_time() and enable_wait_long().

pub fn disable_wait(&mut self) -> Result<(), Error<E>>

Disable the wait feature.

pub fn enable_wait_long(&mut self) -> Result<(), Error<E>>

Enable long wait.

The wait time will be multiplied by 12 so that each cycle takes 0.03s. See also: set_wait_time().

Wait must be enabled with enable_wait().

pub fn disable_wait_long(&mut self) -> Result<(), Error<E>>

Disable long wait.

pub fn set_wait_time(&mut self, value: u8) -> Result<(), Error<E>>

Set the waiting time between proximity and / or color and ambient light cycles.

The value parameter must be a 2’s complement of the number of cycles.

Per default this is set to 0xFF (1 cycle) and each cycle has a fixed duration of 2.78ms except if long wait is enabled, then this time is multiplied by 12.

This must be set before enabling proximity and / or color and ambient light detection.

Waiting must be enabled with enable_wait(). Long wait can be enabled with enable_wait_long().

pub fn force_interrupt(&mut self) -> Result<(), Error<E>>

Force an interrupt.

pub fn clear_interrupts(&mut self) -> Result<(), Error<E>>

Clear all non-gesture interrupts.

impl<I2C, E> Apds9960<I2C, Async>
where I2C: I2cAsync<Error = E>,

Common configuration.

pub async fn enable(&mut self) -> Result<(), Error<E>>

Turn power on.

pub async fn disable(&mut self) -> Result<(), Error<E>>

Deactivate everything and put the device to sleep.

pub async fn enable_wait(&mut self) -> Result<(), Error<E>>

Enable the wait feature.

Enables delay between proximity and / or color and ambient light cycles. The duration of the wait can be configured with set_wait_time() and enable_wait_long().

pub async fn disable_wait(&mut self) -> Result<(), Error<E>>

Disable the wait feature.

pub async fn enable_wait_long(&mut self) -> Result<(), Error<E>>

Enable long wait.

The wait time will be multiplied by 12 so that each cycle takes 0.03s. See also: set_wait_time().

Wait must be enabled with enable_wait().

pub async fn disable_wait_long(&mut self) -> Result<(), Error<E>>

Disable long wait.

pub async fn set_wait_time(&mut self, value: u8) -> Result<(), Error<E>>

Set the waiting time between proximity and / or color and ambient light cycles.

The value parameter must be a 2’s complement of the number of cycles.

Per default this is set to 0xFF (1 cycle) and each cycle has a fixed duration of 2.78ms except if long wait is enabled, then this time is multiplied by 12.

This must be set before enabling proximity and / or color and ambient light detection.

Waiting must be enabled with enable_wait(). Long wait can be enabled with enable_wait_long().

pub async fn force_interrupt(&mut self) -> Result<(), Error<E>>

Force an interrupt.

pub async fn clear_interrupts(&mut self) -> Result<(), Error<E>>

Clear all non-gesture interrupts.

impl<I2C, E> Apds9960<I2C, SyncNonBlocking>
where I2C: I2c<Error = E>,

Gesture engine configuration.

pub fn enable_gesture(&mut self) -> Result<(), Error<E>>

Enable gesture detection

pub fn disable_gesture(&mut self) -> Result<(), Error<E>>

Disable gesture detection

pub fn enable_gesture_mode(&mut self) -> Result<(), Error<E>>

Enable gesture mode.

This can be automatically enabled (depending on proximity thresholds) and disabled (see GMODE on datasheet).

pub fn disable_gesture_mode(&mut self) -> Result<(), Error<E>>

Disable gesture mode.

This can be automatically enabled (depending on proximity thresholds) and disabled (see GMODE on datasheet).

pub fn enable_gesture_interrupts(&mut self) -> Result<(), Error<E>>

Enable gesture interrupt generation

pub fn disable_gesture_interrupts(&mut self) -> Result<(), Error<E>>

Disable gesture interrupt generation

pub fn set_gesture_data_level_threshold( &mut self, threshold: GestureDataThreshold, ) -> Result<(), Error<E>>

Set the threshold of amount of available data in the gesture FIFO registers.

pub fn set_gesture_proximity_entry_threshold( &mut self, threshold: u8, ) -> Result<(), Error<E>>

Set the gesture proximity entry threshold.

pub fn set_gesture_proximity_exit_threshold( &mut self, threshold: u8, ) -> Result<(), Error<E>>

Set the gesture proximity exit threshold.

pub fn set_gesture_up_offset(&mut self, offset: i8) -> Result<(), Error<E>>

Set the gesture up offset.

pub fn set_gesture_down_offset(&mut self, offset: i8) -> Result<(), Error<E>>

Set the gesture down offset.

pub fn set_gesture_left_offset(&mut self, offset: i8) -> Result<(), Error<E>>

Set the gesture left offset.

pub fn set_gesture_right_offset(&mut self, offset: i8) -> Result<(), Error<E>>

Set the gesture right offset.

pub fn set_gesture_offsets( &mut self, offset_up: i8, offset_down: i8, offset_left: i8, offset_right: i8, ) -> Result<(), Error<E>>

Set the gesture up, down, left and right offsets.

impl<I2C, E> Apds9960<I2C, Async>
where I2C: I2cAsync<Error = E>,

Gesture engine configuration.

pub async fn enable_gesture(&mut self) -> Result<(), Error<E>>

Enable gesture detection

pub async fn disable_gesture(&mut self) -> Result<(), Error<E>>

Disable gesture detection

pub async fn enable_gesture_mode(&mut self) -> Result<(), Error<E>>

Enable gesture mode.

This can be automatically enabled (depending on proximity thresholds) and disabled (see GMODE on datasheet).

pub async fn disable_gesture_mode(&mut self) -> Result<(), Error<E>>

Disable gesture mode.

This can be automatically enabled (depending on proximity thresholds) and disabled (see GMODE on datasheet).

pub async fn enable_gesture_interrupts(&mut self) -> Result<(), Error<E>>

Enable gesture interrupt generation

pub async fn disable_gesture_interrupts(&mut self) -> Result<(), Error<E>>

Disable gesture interrupt generation

pub async fn set_gesture_data_level_threshold( &mut self, threshold: GestureDataThreshold, ) -> Result<(), Error<E>>

Set the threshold of amount of available data in the gesture FIFO registers.

pub async fn set_gesture_proximity_entry_threshold( &mut self, threshold: u8, ) -> Result<(), Error<E>>

Set the gesture proximity entry threshold.

pub async fn set_gesture_proximity_exit_threshold( &mut self, threshold: u8, ) -> Result<(), Error<E>>

Set the gesture proximity exit threshold.

pub async fn set_gesture_up_offset( &mut self, offset: i8, ) -> Result<(), Error<E>>

Set the gesture up offset.

pub async fn set_gesture_down_offset( &mut self, offset: i8, ) -> Result<(), Error<E>>

Set the gesture down offset.

pub async fn set_gesture_left_offset( &mut self, offset: i8, ) -> Result<(), Error<E>>

Set the gesture left offset.

pub async fn set_gesture_right_offset( &mut self, offset: i8, ) -> Result<(), Error<E>>

Set the gesture right offset.

pub async fn set_gesture_offsets( &mut self, offset_up: i8, offset_down: i8, offset_left: i8, offset_right: i8, ) -> Result<(), Error<E>>

Set the gesture up, down, left and right offsets.

impl<I2C, E> Apds9960<I2C, SyncNonBlocking>
where I2C: I2c<Error = E>,

Gesture data reading.

pub fn read_gesture_data_level(&mut self) -> Result<u8, Error<E>>

Read the amount of available data in the gesture FIFO registers.

pub fn is_gesture_data_valid(&mut self) -> Result<bool, Error<E>>

Read whether there is valid gesture data available.

pub fn has_gesture_data_overflown(&mut self) -> Result<bool, Error<E>>

Read whether the gesture data has overflown.

pub fn read_gesture_data(&mut self, data: &mut [u8]) -> Result<(), Error<E>>

Read gesture data.

Will read the gesture data up to the minimum of: gesture data level, array size. Make sure to provide an array with at least the number of elements returned by the read_gesture_data_level() method multiplied by 4.

The data contents will be organized as follows: [up_dataset0, down_dataset0, left_dataset0, right_dataset0, up_dataset1, down_dataset1, left_dataset1, right_dataset1, ...]

Blocks as long as not enough data is available.

impl<I2C, E> Apds9960<I2C, Async>
where I2C: I2cAsync<Error = E>,

Gesture data reading.

pub async fn read_gesture_data_level(&mut self) -> Result<u8, Error<E>>

Read the amount of available data in the gesture FIFO registers.

pub async fn is_gesture_data_valid(&mut self) -> Result<bool, Error<E>>

Read whether there is valid gesture data available.

pub async fn has_gesture_data_overflown(&mut self) -> Result<bool, Error<E>>

Read whether the gesture data has overflown.

pub async fn read_gesture_data( &mut self, data: &mut [u8], ) -> Result<(), Error<E>>

Read gesture data.

Will read the gesture data up to the minimum of: gesture data level, array size. Make sure to provide an array with at least the number of elements returned by the read_gesture_data_level() method multiplied by 4.

The data contents will be organized as follows: [up_dataset0, down_dataset0, left_dataset0, right_dataset0, up_dataset1, down_dataset1, left_dataset1, right_dataset1, ...]

Blocks as long as not enough data is available.

impl<I2C, E> Apds9960<I2C, SyncNonBlocking>
where I2C: I2c<Error = E>,

Color and ambient light.

pub fn enable_light(&mut self) -> Result<(), Error<E>>

Enable color and ambient light detection.

pub fn disable_light(&mut self) -> Result<(), Error<E>>

Disable color and ambient light detection.

pub fn set_light_integration_time(&mut self, value: u8) -> Result<(), Error<E>>

Set the color and ambient light integration time.

The value parameter must be a 2’s complement of the number of cycles.

Per default this is set to 0xFF (1 cycle) and each cycle has a fixed duration of 2.78ms.

pub fn enable_light_interrupts(&mut self) -> Result<(), Error<E>>

Enable ambient light interrupt generation.

pub fn disable_light_interrupts(&mut self) -> Result<(), Error<E>>

Disable ambient light interrupt generation.

pub fn enable_light_saturation_interrupts(&mut self) -> Result<(), Error<E>>

Enable clear channel ambient light saturation interrupt generation.

pub fn disable_light_saturation_interrupts(&mut self) -> Result<(), Error<E>>

Disable clear channel ambient light saturation interrupt generation.

pub fn set_light_low_threshold( &mut self, threshold: u16, ) -> Result<(), Error<E>>

Set the clear channel ambient light interrupt low threshold.

An interrupt will be generated if light interrupts are enabled and the clear data is less than this value.

pub fn set_light_high_threshold( &mut self, threshold: u16, ) -> Result<(), Error<E>>

Set the clear channel ambient light interrupt high threshold.

An interrupt will be generated if light interrupts are enabled and the clear data is greater than this value.

pub fn clear_light_interrupt(&mut self) -> Result<(), Error<E>>

Clear ambient light interrupt.

pub fn read_light(&mut self) -> NbResult<LightData, Error<E>>

Read the color / ambient light sensor data.

Returns nb::Error::WouldBlock as long as the data is not ready. This clears the data ready flag.

pub fn read_light_clear(&mut self) -> NbResult<u16, Error<E>>

Read the color / ambient light sensor clear channel data.

Blocks (asynchronously) or returns nb::Error::WouldBlock when the data is not ready. This clears the data ready flag.

pub fn read_light_red(&mut self) -> NbResult<u16, Error<E>>

Read the color / ambient light sensor red channel data.

Blocks (asynchronously) or returns nb::Error::WouldBlock when the data is not ready. This clears the data ready flag.

pub fn read_light_green(&mut self) -> NbResult<u16, Error<E>>

Read the color / ambient light sensor green channel data.

Blocks (asynchronously) or returns nb::Error::WouldBlock when the data is not ready. This clears the data ready flag.

pub fn read_light_blue(&mut self) -> NbResult<u16, Error<E>>

Read the color / ambient light sensor blue channel data.

Blocks (asynchronously) or returns nb::Error::WouldBlock when the data is not ready. This clears the data ready flag.

pub fn is_light_data_valid(&mut self) -> Result<bool, Error<E>>

Read whether the color and ambient light sensor data is valid.

This is checked internally in the read_light_*() methods as well.

impl<I2C, E> Apds9960<I2C, Async>
where I2C: I2cAsync<Error = E>,

Color and ambient light.

pub async fn enable_light(&mut self) -> Result<(), Error<E>>

Enable color and ambient light detection.

pub async fn disable_light(&mut self) -> Result<(), Error<E>>

Disable color and ambient light detection.

pub async fn set_light_integration_time( &mut self, value: u8, ) -> Result<(), Error<E>>

Set the color and ambient light integration time.

The value parameter must be a 2’s complement of the number of cycles.

Per default this is set to 0xFF (1 cycle) and each cycle has a fixed duration of 2.78ms.

pub async fn enable_light_interrupts(&mut self) -> Result<(), Error<E>>

Enable ambient light interrupt generation.

pub async fn disable_light_interrupts(&mut self) -> Result<(), Error<E>>

Disable ambient light interrupt generation.

pub async fn enable_light_saturation_interrupts( &mut self, ) -> Result<(), Error<E>>

Enable clear channel ambient light saturation interrupt generation.

pub async fn disable_light_saturation_interrupts( &mut self, ) -> Result<(), Error<E>>

Disable clear channel ambient light saturation interrupt generation.

pub async fn set_light_low_threshold( &mut self, threshold: u16, ) -> Result<(), Error<E>>

Set the clear channel ambient light interrupt low threshold.

An interrupt will be generated if light interrupts are enabled and the clear data is less than this value.

pub async fn set_light_high_threshold( &mut self, threshold: u16, ) -> Result<(), Error<E>>

Set the clear channel ambient light interrupt high threshold.

An interrupt will be generated if light interrupts are enabled and the clear data is greater than this value.

pub async fn clear_light_interrupt(&mut self) -> Result<(), Error<E>>

Clear ambient light interrupt.

pub async fn read_light(&mut self) -> Result<LightData, Error<E>>

Read the color / ambient light sensor data.

Blocks as long as the data is not ready. This clears the data ready flag.

pub async fn read_light_clear(&mut self) -> Result<u16, Error<E>>

Read the color / ambient light sensor clear channel data.

Blocks (asynchronously) or returns nb::Error::WouldBlock when the data is not ready. This clears the data ready flag.

pub async fn read_light_red(&mut self) -> Result<u16, Error<E>>

Read the color / ambient light sensor red channel data.

Blocks (asynchronously) or returns nb::Error::WouldBlock when the data is not ready. This clears the data ready flag.

pub async fn read_light_green(&mut self) -> Result<u16, Error<E>>

Read the color / ambient light sensor green channel data.

Blocks (asynchronously) or returns nb::Error::WouldBlock when the data is not ready. This clears the data ready flag.

pub async fn read_light_blue(&mut self) -> Result<u16, Error<E>>

Read the color / ambient light sensor blue channel data.

Blocks (asynchronously) or returns nb::Error::WouldBlock when the data is not ready. This clears the data ready flag.

pub async fn is_light_data_valid(&mut self) -> Result<bool, Error<E>>

Read whether the color and ambient light sensor data is valid.

This is checked internally in the read_light_*() methods as well.

impl<I2C, E> Apds9960<I2C, SyncNonBlocking>
where I2C: I2c<Error = E>,

Proximity.

pub fn enable_proximity(&mut self) -> Result<(), Error<E>>

Enable proximity detection

pub fn disable_proximity(&mut self) -> Result<(), Error<E>>

Disable proximity detection

pub fn enable_proximity_interrupts(&mut self) -> Result<(), Error<E>>

Enable proximity interrupt generation

pub fn disable_proximity_interrupts(&mut self) -> Result<(), Error<E>>

Disable proximity interrupt generation

pub fn enable_proximity_saturation_interrupts(&mut self) -> Result<(), Error<E>>

Enable proximity saturation interrupt generation

pub fn disable_proximity_saturation_interrupts( &mut self, ) -> Result<(), Error<E>>

Disable proximity saturation interrupt generation

pub fn set_proximity_low_threshold( &mut self, threshold: u8, ) -> Result<(), Error<E>>

Set the proximity interrupt low threshold.

pub fn set_proximity_high_threshold( &mut self, threshold: u8, ) -> Result<(), Error<E>>

Set the proximity interrupt high threshold.

pub fn set_proximity_up_right_offset( &mut self, offset: i8, ) -> Result<(), Error<E>>

Set the proximity up/right photodiode offset.

pub fn set_proximity_down_left_offset( &mut self, offset: i8, ) -> Result<(), Error<E>>

Set the proximity down/left photodiode offset.

pub fn set_proximity_offsets( &mut self, offset_up_right: i8, offset_down_left: i8, ) -> Result<(), Error<E>>

Set the proximity up/right and down/left photodiode offset.

pub fn clear_proximity_interrupt(&mut self) -> Result<(), Error<E>>

Clear proximity interrupt.

pub fn read_proximity(&mut self) -> Result<u8, Error<E>>

Read the proximity sensor data.

Blocks as long as the data is not ready.

pub fn is_proximity_data_valid(&mut self) -> Result<bool, Error<E>>

Read whether the proximity sensor data is valid.

This is checked internally in read_proximity() as well.

impl<I2C, E> Apds9960<I2C, Async>
where I2C: I2cAsync<Error = E>,

Proximity.

pub async fn enable_proximity(&mut self) -> Result<(), Error<E>>

Enable proximity detection

pub async fn disable_proximity(&mut self) -> Result<(), Error<E>>

Disable proximity detection

pub async fn enable_proximity_interrupts(&mut self) -> Result<(), Error<E>>

Enable proximity interrupt generation

pub async fn disable_proximity_interrupts(&mut self) -> Result<(), Error<E>>

Disable proximity interrupt generation

pub async fn enable_proximity_saturation_interrupts( &mut self, ) -> Result<(), Error<E>>

Enable proximity saturation interrupt generation

pub async fn disable_proximity_saturation_interrupts( &mut self, ) -> Result<(), Error<E>>

Disable proximity saturation interrupt generation

pub async fn set_proximity_low_threshold( &mut self, threshold: u8, ) -> Result<(), Error<E>>

Set the proximity interrupt low threshold.

pub async fn set_proximity_high_threshold( &mut self, threshold: u8, ) -> Result<(), Error<E>>

Set the proximity interrupt high threshold.

pub async fn set_proximity_up_right_offset( &mut self, offset: i8, ) -> Result<(), Error<E>>

Set the proximity up/right photodiode offset.

pub async fn set_proximity_down_left_offset( &mut self, offset: i8, ) -> Result<(), Error<E>>

Set the proximity down/left photodiode offset.

pub async fn set_proximity_offsets( &mut self, offset_up_right: i8, offset_down_left: i8, ) -> Result<(), Error<E>>

Set the proximity up/right and down/left photodiode offset.

pub async fn clear_proximity_interrupt(&mut self) -> Result<(), Error<E>>

Clear proximity interrupt.

pub async fn read_proximity(&mut self) -> Result<u8, Error<E>>

Read the proximity sensor data.

Blocks as long as the data is not ready.

pub async fn is_proximity_data_valid(&mut self) -> Result<bool, Error<E>>

Read whether the proximity sensor data is valid.

This is checked internally in read_proximity() as well.

impl<I2C, E> Apds9960<I2C, SyncNonBlocking>
where I2C: I2c<Error = E>,

pub fn read_device_id(&mut self) -> Result<u8, Error<E>>

Read the device ID.

This is per default 0xAB.

impl<I2C, E> Apds9960<I2C, Async>
where I2C: I2cAsync<Error = E>,

pub async fn read_device_id(&mut self) -> Result<u8, Error<E>>

Read the device ID.

This is per default 0xAB.

impl<I2C, M: Mode> Apds9960<I2C, M>

pub fn new(i2c: I2C) -> Self

Create new instance of the APDS9960 device.

pub fn destroy(self) -> I2C

Destroy driver instance, return I²C bus instance.

impl<I2C> Apds9960<I2C, SyncNonBlocking>

pub fn new_non_blocking(i2c: I2C) -> Self

Create new instance of the APDS9960 device.

impl<I2C> Apds9960<I2C, Async>

pub fn new_async(i2c: I2C) -> Self

Create new instance of the APDS9960 device.